Refactored the storyteller microservice - Added environment variables instead of hardcode> - Beautified variable names and fstrings <abbitmq connection
