+++ storyteller/app/internal/__init__.py
@@ -1,3 +1,3 @@
-from .lines import Story
+from .story import Story
 from .rabbitmq import RabbitMQConnection
 from .edenai import EdenAI
