+++ storyteller/app/internal/edenai.py
@@ -3,8 +3,6 @@ import requests
 import os
 
 HEADERS = {"Authorization": "Bearer " + os.getenv("EDENAI_TOKEN")} 
-URL = "https://api.edenai.run/v2/text/generation"
-PROVIDER = "openai"
 
 class EdenAI():
     def __init__(self, text: str, lines: int):
@@ -12,9 +10,9 @@ class EdenAI():
         self.lines = lines
     def run(self):
         payload = {
-            "providers": PROVIDER,
-            "text": "Can you please write a " + str(self.lines) +" line story regarding " + self.text + "?",
+            "providers": os.getenv("EDENAI_PROVIDER"),
+            "text": f"Can you please write a {self.lines} line story regarding {self.text}?",
             "temperature": 0.2
         }
-        response = requests.post(URL, json=payload, headers=HEADERS)
-        return json.loads(response.text)[PROVIDER]['generated_text']
+        response = requests.post(os.getenv("EDENAI_URL"), json=payload, headers=HEADERS)
+        return json.loads(response.text)[os.getenv("EDENAI_PROVIDER")]['generated_text']
