+++ storyteller/app/main.py
@@ -1,9 +1,17 @@
 from fastapi import FastAPI
+from contextlib import asynccontextmanager
 import uvicorn
+import os
+from internal import RabbitMQConnection
 from routers import router
 
-app = FastAPI()
-app.include_router(router)
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    app.rabbitmq_connection = RabbitMQConnection(host=os.getenv("RABBIT_HOST"), username=os.getenv('username'), password=os.getenv('password'))
+    
+    yield
+    
+    app.rabbitmq_connection.connection.close()
 
-if __name__ == "__main__":
-   uvicorn.run("main:app", host="127.0.0.1", port=8000, reload=True)
+app = FastAPI(lifespan=lifespan)
+app.include_router(router)
